requirements.txt
streamlit==1.33.0
openai==1.21.1
langchain==0.1.20
python-dotenv
PyMuPDF
langchain-openai==0.1.6